refactor: split commands code into separate files
